fix state reset of creating a event when keyboard appears. added validation on empty
